+++ lib/src/errors/error_of.dart
@@ -20,7 +20,7 @@ class ErrorOf<T> extends Error {
   final Object expectedState;
 
   /// Whether to enable color output.
-  static ColorOutput colorOutput = ColorOutput.ON;
+  static ColorOutput colorOutput = ColorOutput.on;
 
   /// Type argument of the error class.
   Type get typeArgument => T;