+++ lib/src/exceptions/exception_of_type.dart
@@ -30,7 +30,7 @@ class ExceptionOfType<T extends ExceptionType> implements Exception {
   final Object expectedState;
 
   /// Whether to enable color output.
-  static ColorOutput colorOutput = ColorOutput.ON;
+  static ColorOutput colorOutput = ColorOutput.on;
 
   /// Type argument of the exception class.
   Type get typeArgument => T;